refactor(#30): Change entities for interfaces
