--- src/entities/matchedTrick.ts
@@ -1,6 +1,0 @@
-import { Trick } from './trick';
-
-export class MatchedTrick extends Trick {
-    captured: string[];
-    element: HTMLElement;
-}
