+++ src/interfaces/matched-trick.interface.ts
@@ -1,0 +1,6 @@
+import { Trick } from './trick.interface';
+
+export interface MatchedTrick extends Trick {
+    captured: string[];
+    element: HTMLElement;
+}
